+++ dash_bio/_imports_.py
@@ -1,5 +1,6 @@
 from .AlignmentChart import AlignmentChart
 from .Circos import Circos
+from .FornaContainer import FornaContainer
 from .Ideogram import Ideogram
 from .Molecule2dViewer import Molecule2dViewer
 from .Molecule3dViewer import Molecule3dViewer
@@ -11,6 +12,7 @@ from .Speck import Speck
 __all__ = [
     "AlignmentChart",
     "Circos",
+    "FornaContainer",
     "Ideogram",
     "Molecule2dViewer",
     "Molecule3dViewer",