+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+* Added FornaContainer component for RNA secondary structure analysis.
+
 ## [0.2.0] - 2019-10-18
 
 ### Fixed